bug fixes on item permanance
Just some minor bugs needed to be fixed, im sure more will apear as we go. time fixed a bug where items wouldnt check if the player picked them up because it was instantiating with the suffex "(clone)" so just added that suffex to the check,  the item names were not displaying because i needed to put a count in the for each loop and i changed the way globitems load items works, so it outputs an item array that i add to the player items now.