updated Dialogue system
AHGh this system works so much better then the last one. so there is now only one dialogue method, the npc has more control over what dialogue is sent to it, on  top of this, the dialouge system now just creates the buttons via instantiation instead of having a hard limitito the amout of buttons easy.