Persistant item equips and convertstringtoitem
So this time it saves the items you have equiped, as strings, then as items.

i wrote a usefull function in globalitemhandler called convertstringtoitem which takes any string, and asllong as the item is in the master list, turns that string into an item.

also clear inventory when you save, and added the string list to the clear inventory function so firstly when you load inside a session without quiting, items dont duplicate and secondaly making it so the string scriptable object isnt filing up with items which are never cleared and cycling back into the game as item dupes in the save load function